refactor: Clean up file structure and remove duplicates
🧹 Structure Cleanup:
- Removed duplicate test-utils files (kept more comprehensive version)
- Removed duplicate config test files (kept more comprehensive version)
- Removed duplicate cn.test.ts files (kept more comprehensive version)
- Moved test files to proper locations (components tests to __tests__/components/)
- Removed empty app/api directory
- Removed test chat route that conflicted with main chat interface
- Removed misplaced .code-workspace file from lib/supabase/

✅ Best Practices Applied:
- Consolidated duplicate files following DRY principle
- Organized test files in proper directory structure
- Removed conflicting routes and empty directories
- Maintained separation of concerns (client vs server auth files)

🔧 Build Status:
- TypeScript compilation passes
- Production build successful
- All functionality preserved
- Error handling working correctly
