+++ app/(chat)/api/document/route.ts
@@ -68,9 +68,9 @@ export async function POST(req: Request) {
 			const { error } = await supabase
 				.from("documents")
 				.update({
-					content: content as any,
-					title: title as any,
-				})
+					content,
+					title,
+				} as any)
 				.eq("id", id)
 				.eq("user_id", user.id);
 
@@ -116,7 +116,7 @@ export async function PATCH(request: Request) {
 
 		const { error } = await supabase
 			.from("documents")
-			.update({ content: content as any, title: title as any })
+			.update({ content, title } as any)
 			.eq("id", id)
 			.eq("user_id", user.id);
