--- app/chat/route.ts
@@ -1,5 +1,0 @@
-import { NextResponse } from 'next/server';
-
-export async function GET() {
-  return NextResponse.json({ message: "Hello" });
-}
