--- lib/supabase/ai-bot-vercel.code-workspace
@@ -1,7 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "../.."
-		}
-	]
-}
